refactor: 삼각형 넓이 계산 시 삼각형 객체를 활용하도록 변경
